Make "webpacker:yarn_install" task env aware (#1331)
* Make webpacker:yarn_install task env aware

When webpacker compiles in a non-production environment it's removing
node_modules which are dev dependencies. This is problematic on CI for
example because these dependencies may be required.

* Use alternative to Bundler.with_clean_env

This was failing on CI presumably because BUNDLE_GEMFILE was being
lost, causing the 4.2.x Gemfile tests to fail.

* Remove explicit production flag from Yarn Install task

If the NODE_ENV is set then Yarn will behave accordingly (e.g. skip
devDependencies if NODE_ENV=production).

Based on discussion:
https://github.com/rails/webpacker/pull/1331#issuecomment-443518808

* Use existing with_node_env helper