Remove success message on webpacker:install
Currently, the last line of output when running `rails new` is
"Webpacker successfully installed" which is somewhat confusing if you
don't read all of the intermediate output from that command. After
proposing https://github.com/rails/rails/pull/42546 to add a clear
success message to the end of `rails new`, @rafaelfranca and others
recommended removing the success message from webpacker to be more in
line the Unix philosophy of not saying anything when things go well.